Entity Framework added
The project uses a SQLite database. There are no migrations as SQLite does not support the full feature set of SQL.
There is an Initialization method that adds working data to the database if it is empty. This will need to be updated as new entities are added to the database.